loops

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	int a = 1, b = 2, c, d;
+	long int a = 1, b = 2, c, d;
 
 	printf("%ld, %ld, ", a, b);
 	for (c = 1; c <= 48; c++)
